requirements.txt
Flask
Flask-Cors
Flask-Limiter
PyPDF2
python-dotenv
boto3
chromadb
langchain
redis
pandas
openpyxl
sentence-transformers
Pillow

